1. 清除重复的库依赖，代码与重复使用的内容整合到common中
