+++ common/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 # 用于处理颜色字符串
 color-string = "0.1.6"
 # 用于序列化和反序列化数据
-bincode = "2.0.0-rc.3"
+#bincode = "2.0.0-rc.3"
 # 用于处理 JSON 数据
 serde_json = "1.0.124"
 # 用于解析和生成 TOML 格式配置文件
@@ -24,3 +24,7 @@ once_cell = "1.19.0"
 chrono = "0.4.38"
 # 线程池
 threadpool = "1.8.1"
+# 用于计算表达式
+evalexpr = "11.3.0"
+# 用于正则表达式
+regex = "1"
