+++ engine/src/core/flow/dispatch/dispatch_loop.rs
@@ -1,10 +1,9 @@
-use crate::core::expression::interface::preliminary_analysis_string;
 use crate::core::flow::dispatch::common::match_node_id;
 use crate::core::flow::dispatch::interface::dispatch_nodes;
 use engine_common::entity::error::{DispatchErr, NodeError};
 use engine_common::entity::flow::{Blueprint, FlowData, Node};
+use engine_common::expr::interface::expr_eval_bool;
 use engine_common::logger::interface::{debug, fail};
-use evalexpr::eval_boolean;
 
 pub async fn dispatch_loop(blueprint: Blueprint, node: Node, flow_data: &mut FlowData) -> Result<(), DispatchErr> {
     let node_id = node.clone().id.expect("node id must be set");
@@ -72,9 +71,8 @@ pub async fn dispatch_loop(blueprint: Blueprint, node: Node, flow_data: &mut Flo
     }
 
 
-    let expr = preliminary_analysis_string(expression.as_str().unwrap().to_string(), flow_data.clone().params);
     // 执行表达式
-    let mut result = match eval_boolean(expr.as_str()) {
+    let mut result = match expr_eval_bool(expression.as_str().unwrap(),flow_data.clone().params) {
         Ok(value) => { value }
         Err(err) => {
             // 表达式执行出错
@@ -95,8 +93,7 @@ pub async fn dispatch_loop(blueprint: Blueprint, node: Node, flow_data: &mut Flo
             debug(format!("Loop main struct start : {}, now in {}", node_id, index + 1).as_str());
             index += 1;
             // 执行表达式
-            let expr = preliminary_analysis_string(expression.as_str().unwrap().to_string(), flow_data.clone().params);
-            result = match eval_boolean(expr.as_str()) {
+            result = match expr_eval_bool(expression.as_str().unwrap(), flow_data.clone().params) {
                 Ok(value) => { value }
                 Err(err) => {
                     // 表达式执行出错
