+++ engine/src/core/flow/dispatch/mod.rs
@@ -2,4 +2,3 @@ pub mod interface;
 mod dispatch_loop;
 mod dispatch_general;
 mod common;
-mod exception;
