+++ handler/Cargo.toml
@@ -12,15 +12,9 @@ serde = { version = "1.0.207", features = ["derive"] }
 serde_derive = "1.0.207"
 # 用于处理 JSON 数据
 serde_json = "1.0.124"
-# 用于序列化和反序列化数据
-bincode = "2.0.0-rc.3"
-# Web 框架
-rocket = { version = "^0.5.1", features = ["json"] }
 # 异步运行时库
 tokio = { version = "1.39.2", features = ["full"] }
 # 加载dll
 libloader = "0.1.4"
 # 加载 so和dylib
 libloading = "0.8.5"
-
-evalexpr = "11.3.0"
